Updated Game to allow unique win/loss states
Added new booleans to determine if win or loss state is active. 
Added a 2 methods to allow these variables to be updated.
Added an abstract method checkVictory() that happens after every computer turn. This is meant to decide whether or not the game should end and can be custom to each game type.